frontend: Remove circular dep in LazyLoadComment styles.
