+++ packages/frontend-web/src/app/components/LazyCommentList/LazyCommentList.tsx
@@ -33,25 +33,17 @@ import { SortColumn } from './components/SortColumn';
 
 import {
   ARTICLE_CAPTION_TYPE,
-  ARTICLE_CATEGORY_TYPE,
   BASE_Z_INDEX,
-  BODY_TEXT_TYPE,
   BOX_DEFAULT_SPACING,
-  BUTTON_LINK_TYPE,
-  BUTTON_RESET,
-  CAPTION_TYPE,
-  DARK_PRIMARY_TEXT_COLOR,
-  DARK_SECONDARY_TEXT_COLOR,
-  DARK_TERTIARY_TEXT_COLOR,
   GUTTER_DEFAULT_SPACING,
   MEDIUM_COLOR,
   OFFSCREEN,
   SELECT_ELEMENT,
   SELECT_Z_INDEX,
 } from '../../styles';
+import { COMMENT_HEADER_BACKGROUND_COLOR, ROW_STYLES } from '../styles';
 
 const ARROW_SIZE = 6;
-const COMMENT_HEADER_BACKGROUND_COLOR = '#F5F7F9';
 const COMMENT_HEADER_HEIGHT = 56;
 const BASE_ROW_HEIGHT = 150;
 const ROW_PADDING = 64;
@@ -110,101 +102,6 @@ const HEADER_STYLES = stylesheet({
 
   label: {
     cursor: 'pointer',
-  },
-});
-
-export const ROW_STYLES = stylesheet({
-  meta: {
-    ...CAPTION_TYPE,
-    color: DARK_SECONDARY_TEXT_COLOR,
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    height: '46px',
-  },
-
-  authorRow: {
-    display: 'flex',
-    flex: 1,
-    alignItems: 'center',
-  },
-
-  commentContainer: {
-    display: 'flex',
-    flex: 1,
-    flexDirection: 'column',
-    paddingBottom: `${GUTTER_DEFAULT_SPACING / 2}px`,
-  },
-
-  comment: {
-    ...BODY_TEXT_TYPE,
-    color: DARK_PRIMARY_TEXT_COLOR,
-    paddingRight: `${GUTTER_DEFAULT_SPACING}px`,
-  },
-
-  commentHeading: {
-    ...BODY_TEXT_TYPE,
-    color: DARK_PRIMARY_TEXT_COLOR,
-    paddingRight: `${GUTTER_DEFAULT_SPACING}px`,
-  },
-
-  reply: {
-    fill: DARK_TERTIARY_TEXT_COLOR,
-    marginBottom: 3,
-    marginRight: 4,
-    verticalAlign: 'top',
-  },
-
-  approval: {
-    ...ARTICLE_CATEGORY_TYPE,
-    color: MEDIUM_COLOR,
-    padding: `0 0 0 ${GUTTER_DEFAULT_SPACING * 1.5}px`,
-    textAlign: 'left',
-  },
-
-  actionToggle: {
-    ...BUTTON_RESET,
-    padding: `${GUTTER_DEFAULT_SPACING / 2}px`,
-    marginRight: `${GUTTER_DEFAULT_SPACING / 4}px`,
-    ':focus': {
-      outline: 0,
-      backgroundColor: `${COMMENT_HEADER_BACKGROUND_COLOR}`,
-    },
-  },
-
-  articleLink: {
-    color: DARK_PRIMARY_TEXT_COLOR,
-    ':focus': {
-      outline: 0,
-      textDecoration: 'underline',
-    },
-  },
-
-  detailsButton: {
-    ...BUTTON_LINK_TYPE,
-    flex: 1,
-    border: 'none',
-    borderRadius: 0,
-    color: MEDIUM_COLOR,
-    cursor: 'pointer',
-    textAlign: 'right',
-    fontSize: '12px',
-    marginRight: `${GUTTER_DEFAULT_SPACING}px`,
-    ':hover': {
-      textDecoration: 'underline',
-    },
-
-    ':focus': {
-      outline: 0,
-      textDecoration: 'underline',
-    },
-  },
-
-  actionContainer: {
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
   },
 });
